Add support for Amazon Linux
Closes #250
