+++ agent/__init__.py
@@ -10,26 +10,21 @@ import glob
 import logging
 import logging.config
 from math import floor
-from sys import exit
-from sys import stdout
+from sys import stdout, exit
 from pathlib import Path
 
 import requests
 import pkg_resources
 from cryptography import x509
 from cryptography.hazmat.backends import default_backend
-from cryptography.hazmat.primitives import hashes
-from cryptography.hazmat.primitives import serialization
+from cryptography.hazmat.primitives import hashes, serialization
 from cryptography.hazmat.primitives.asymmetric import ec
 from cryptography.x509.oid import NameOID
 import pytz
 
-from agent import iptables_helper
-from agent import journal_helper
-from agent import rpi_helper
-from agent import security_helper
+from agent import iptables_helper, journal_helper, os_helper, security_helper
 from agent.executor import Locker
-from agent.rpi_helper import Confinement, detect_cloud, detect_confinement, detect_installation, get_deb_packages
+from agent.os_helper import Confinement, detect_confinement, detect_cloud, detect_installation, get_packages
 
 
 CONFINEMENT = detect_confinement()
@@ -112,7 +107,7 @@ def get_primary_ip():
         addrs = netifaces.ifaddresses(primary_interface)
         return addrs[netifaces.AF_INET][0]['addr']
     except (OSError, KeyError):
-        return None
+        return
 
 
 def get_certificate_expiration_date():
@@ -132,7 +127,8 @@ def get_certificate_expiration_date():
 
 def time_for_certificate_renewal():
     """ Check if it's time for certificate renewal """
-    return datetime.datetime.now(datetime.timezone.utc) + datetime.timedelta(days=RENEWAL_THRESHOLD) > get_certificate_expiration_date()
+    return datetime.datetime.now(datetime.timezone.utc) + datetime.timedelta(days=RENEWAL_THRESHOLD) > \
+           get_certificate_expiration_date()
 
 
 def is_certificate_expired():
@@ -393,11 +389,11 @@ def send_ping(dev=False):
         'confinement': CONFINEMENT.name,
         'cloud': detect_cloud().name,
         'installation': detect_installation().name,
-        'os_release': rpi_helper.get_os_release()
+        'os_release': os_helper.get_os_release()
     }
 
     if CONFINEMENT != Confinement.SNAP:
-        packages = get_deb_packages()
+        packages = get_packages()
         if ping.get('deb_packages_hash') != packages['hash']:
             payload['deb_packages'] = packages
 
@@ -420,13 +416,13 @@ def send_ping(dev=False):
         payload.update({
             'default_password_users': security_helper.check_for_default_passwords(CONFIG_PATH),
             'audit_files': security_helper.audit_config_files(),
-            'auto_upgrades': rpi_helper.auto_upgrades_enabled(),
+            'auto_upgrades': os_helper.auto_upgrades_enabled(),
             'mysql_root_access': security_helper.mysql_root_access(),
-            'kernel_package': rpi_helper.kernel_deb_package(),
+            'kernel_package': os_helper.kernel_package(),
             'cpu': security_helper.cpu_vulnerabilities()
         })
 
-    rpi_metadata = rpi_helper.detect_raspberry_pi()
+    rpi_metadata = os_helper.detect_raspberry_pi()
     if rpi_metadata['is_raspberry_pi']:
         payload.update({
             'device_manufacturer': 'Raspberry Pi',
@@ -750,7 +746,8 @@ def run(ping=True, dev=False, logger=logger):
             if not time_for_certificate_renewal() and not is_certificate_expired():
                 if ping:
                     send_ping(dev=dev)
-                    time_to_cert_expires = get_certificate_expiration_date() - datetime.datetime.now(datetime.timezone.utc)
+                    time_to_cert_expires = get_certificate_expiration_date() - datetime.datetime.now(
+                        datetime.timezone.utc)
                     logger.info(
                         "Certificate expires in {} days and {} hours. No need for renewal."
                         "Renewal threshold is set to {} days.".format(
